requirements.txt
Pillow>=10.0.0
requests>=2.31.0
pytest>=7.4.0

